ai-news-digest/requirements.txt
# AI News Digest - Python Dependencies

# RSS Feed parsing
feedparser>=6.0.0

# HTTP requests
requests>=2.28.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Date parsing
python-dateutil>=2.8.0

# Google Gemini API
google-generativeai>=0.3.0

# Testing
pytest>=7.0.0
